add:提供中心控制完善部分代码

+++ localwaf/src/utils/request.ts
@@ -1,12 +1,13 @@
 import axios from 'axios';
 import proxy from '../config/host';
 import router from '../router/index';
-import  {AesDecrypt,AesEncrypt,isObject} from './usuallytool'
+import  {AesDecrypt,AesEncrypt,isObject,isInList} from './usuallytool'
 
 const env = import.meta.env.MODE || 'development';
 
 const API_HOST = env === 'mock' ? '/' : proxy[env].API; // 如果是mock模式 就不配置host 会走本地Mock拦截
 
+const noVisitClientList = ["/center/list", "logout", "public/login"];
 const CODE = {
   LOGIN_TIMEOUT: 1000,
   REQUEST_SUCCESS: 0,
@@ -50,6 +51,16 @@ instance.interceptors.request.use(
       config.headers['X-Token'] = token
       //config.headers.Authorization =  + token
     }
+    //如果有远控机器
+    let remoteBean =localStorage.getItem("current_server")? localStorage.getItem("current_server"):"" //此处换成自己获取回来的token，通常存在在cookie或者store里面
+
+    if (remoteBean && !isInList(config.url,noVisitClientList) ) {
+      console.log(config)
+      remoteBean = JSON.parse(localStorage.getItem("current_server"))
+      // 让每个请求携带token-- ['X-Token']为自定义key 请根据实际情况自行修改
+      config.headers['Remote-Waf-User-Id'] =  remoteBean.client_tenant_id+"@"+remoteBean.client_user_code
+      //config.headers.Authorization =  + token
+    }
     /*if(config.headers['Content-Type'] !=undefined && config.headers['Content-Type']=="application/json" ){
       data = JSON.stringify(config.data)
 
@@ -76,10 +87,20 @@ instance.interceptors.response.use(
 
         //console.log("再加密后",AesEncrypt(tmpSrcContent))
         return data;
-      }else if(data.code === CODE.AUTH_FAILURE){
+      }else {
+        //如果有远控机器
+        let remoteBean =localStorage.getItem("current_server")? localStorage.getItem("current_server"):"" //此处换成自己获取回来的token，通常存在在cookie或者store里面
+
+        if(!remoteBean  && data.code === CODE.AUTH_FAILURE){
           localStorage.clear();     //删除用户信息
           console.log("鉴权失败")
           router.replace({path: '/login'})
+        }else if(remoteBean  && data.code === CODE.AUTH_FAILURE){
+          remoteBean = JSON.parse(localStorage.getItem("current_server"))
+          data.code = -1
+          data.msg = remoteBean.client_server_name + " 远端鉴权失败"
+          console.log("远端鉴权失败")
+        }
       }
       return data;
     }
